feat: improve graph view ux
